Adding all scopes to the implemented functions

+++ app/components/brand/brandController.js
@@ -3,7 +3,9 @@ const Brand = require('./brand');
 const { isValidScope } = require('../../services/utils');
 const component = 'Brand';
 
+
 const getBrands = async (req, res) => {
+    if(!isValidScope(getBrands.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'})
     try {
         const brands = await Brand.getBrands();
         return res
@@ -17,6 +19,7 @@ const getBrands = async (req, res) => {
 }
 
 const getBrandById = async (req, res) => {
+    if(!isValidScope(getBrandById.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'})
     const { id } = req.params;
     try {
         const brand = await Brand.getBrandById(id);
@@ -38,6 +41,7 @@ const getBrandById = async (req, res) => {
 }
 
 const createBrand = async (req, res) => {
+     if(!isValidScope(createBrand.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'});
     const { brandName } = req.body;
     if (!brandName) {
         return res
@@ -58,6 +62,7 @@ const createBrand = async (req, res) => {
 }
 
 const updateBrand = async (req, res) => {
+     if(!isValidScope(updateBrand.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'});
     const { id } = req.params;
     const { body } = req;
     try {
@@ -79,6 +84,7 @@ const updateBrand = async (req, res) => {
 }
 
 const deleteBrand = async (req, res) => {
+     if(!isValidScope(deleteBrand.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'});
     const { id } = req.params;
     try {
         const wasDeleted = await Brand.deleteBrand(id);
