+++ app/components/job/jobController.js
@@ -5,6 +5,7 @@ const { isValidScope } = require('../../services/utils')
 const component = 'Job';
 
 const getJobs = async (req, res) => {
+     if(!isValidScope(getJobs.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'});
     try {
         const jobs = await Job.getJobs();
         if (jobs.length > 0) {
@@ -25,6 +26,7 @@ const getJobs = async (req, res) => {
 }
 
 const getJobById = async (req, res) => {
+     if(!isValidScope(getJobById.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'});
     const { id } = req.params;
     try {
         const job = await Job.getJobById(id);
@@ -46,6 +48,7 @@ const getJobById = async (req, res) => {
 }
 
 const createJob = async (req, res) => {
+     if(!isValidScope(createJob.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'});
     const { body } = req;
     try {
         const job = await Job.createJob(body);
@@ -67,6 +70,7 @@ const createJob = async (req, res) => {
 }
 
 const updateJob = async (req, res) => {
+     if(!isValidScope(updateJob.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'});
     const { id } = req.params;
     const { body } = req;
     try {
@@ -89,6 +93,7 @@ const updateJob = async (req, res) => {
 }
 
 const deleteJob = async (req, res) => {
+     if(!isValidScope(deleteJob.name, component)) return res.status(httpStatus.UNAUTHORIZED).send({ message: 'Usted no cuenta con permisos para ejecutar esta acción'});
     const { id } = req.params;
     try {
         const wasDeleted = await Job.deleteJob(id);
